# openapi.yaml
openapi: 3.0.3
info:
  title: SalesMind API
  description: Mental fitness platform for sales teams - API for managing user profiles, assessments, skills, drills, and progress tracking
  version: 1.0.0
  contact:
    name: SalesMind Support
    email: support@salesmind.com
servers:
  - url: https://api.salesmind.com/v1
    description: Production server
  - url: https://staging-api.salesmind.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /users/profile:
    get:
      tags:
        - User Profiles
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - User Profiles
      summary: Update user profile
      description: Update the authenticated user's profile information
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: Jane Smith
                role:
                  type: string
                  example: Senior Sales Manager
                company:
                  type: string
                  example: Acme Corp
                language_preference:
                  type: string
                  enum: [en, ru]
                  example: en
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /skills:
    get:
      tags:
        - Skills
      summary: List all available skills
      description: Retrieve a list of all mental skills available in the platform
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter skills by category
          schema:
            type: string
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Skills retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
                  count:
                    type: integer
                    example: 14
        '401':
          $ref: '#/components/responses/Unauthorized'
  /skills/{skillId}:
    get:
      tags:
        - Skills
      summary: Get skill details
      description: Retrieve detailed information about a specific skill
      security:
        - BearerAuth: []
      parameters:
        - name: skillId
          in: path
          required: true
          description: Unique identifier of the skill
          schema:
            type: string
            example: emotional-resilience
      responses:
        '200':
          description: Skill details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/skills:
    get:
      tags:
        - Skills
      summary: Get user's skill scores
      description: Retrieve the authenticated user's skill scores and progress
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User skills retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSkill'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/skills/{skillId}:
    get:
      tags:
        - Skills
      summary: Get specific user skill
      description: Retrieve the user's score and progress for a specific skill
      security:
        - BearerAuth: []
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            example: confidence-building
      responses:
        '200':
          description: User skill retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSkill'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Skills
      summary: Update user skill score
      description: Update the score for a specific skill
      security:
        - BearerAuth: []
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_score
              properties:
                current_score:
                  type: integer
                  minimum: 0
                  maximum: 100
                  example: 75
      responses:
        '200':
          description: Skill score updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSkill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or signup
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          example: John Doe
        role:
          type: string
          example: Sales Representative
        company:
          type: string
          example: Tech Solutions Inc
        language_preference:
          type: string
          enum: [en, ru]
          example: en
        onboarding_completed:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Skill:
      type: object
      properties:
        id:
          type: string
          example: emotional-resilience
        name:
          type: string
          example: Emotional Resilience to Rejection
        short_name:
          type: string
          example: Emotional Resilience
        description:
          type: string
          example: Build mental toughness to bounce back quickly from rejection and maintain motivation.
        icon:
          type: string
          example: 🛡️
        methodologies:
          type: array
          items:
            type: string
          example: ["Cognitive Behavioral Therapy (CBT)", "Self-compassion techniques"]
        category:
          type: string
          example: general
        is_active:
          type: boolean
          example: true
    SkillDetail:
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          properties:
            drills:
              type: array
              items:
                $ref: '#/components/schemas/Drill'
            user_score:
              type: integer
              minimum: 0
              maximum: 100
              example: 72
    UserSkill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        skill_id:
          type: string
          example: confidence-building
        skill_name:
          type: string
          example: Confidence Building
        current_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 45
        initial_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 30
        improvement:
          type: integer
          example: 15
        last_updated:
          type: string
          format: date-time
    Drill:
      type: object
      properties:
        id:
          type: string
          example: abc-card
        skill_id:
          type: string
          example: emotional-resilience
        title:
          type: string
          example: ABC Card Analysis
        description:
          type: string
          example: Identify and challenge negative thought patterns after rejection
        type:
          type: string
          enum: [form, timer, interactive, tracking]
          example: form
        component:
          type: string
          example: AbcCardDrill
        duration_minutes:
          type: integer
          example: 10
        difficulty_level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: beginner
        is_active:
          type: boolean
          example: true
        completion_count:
          type: integer
          description: Number of times user has completed this drill
          example: 3
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: The provided data is invalid
        status_code:
          type: integer
          example: 400
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication token is missing or invalid
            status_code: 401
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: The requested resource does not exist
            status_code: 404